SparkCacheConfigurationExtension plus tests
